Update bootstrap + refactor
