--- examples/shopping-list-with-redux-hooks/client/components/Form.js
@@ -1,5 +1,0 @@
-import { connectRootBasedUrls } from 'resolve-redux'
-
-const Form = connectRootBasedUrls(['action'])('form')
-
-export default Form
