+++ examples/shopping-list-with-redux-hooks/client/components/MyLists.js
@@ -10,22 +10,3 @@ export default () => {
     </div>
   )
 }
-
-/*
-export const mapStateToOptions = () => ({
-  readModelName: 'ShoppingLists',
-  resolverName: 'all',
-  resolverArgs: {}
-})
-
-export const mapStateToProps = state => ({
-  lists: state.optimisticShoppingLists || []
-})
-
-export const mapDispatchToProps = dispatch =>
-  bindActionCreators(aggregateActions, dispatch)
-
-export default connectReadModel(mapStateToOptions)(
-  connect(mapStateToProps, mapDispatchToProps)(MyLists)
-)
-*/
