+++ examples/shopping-list-with-redux-hooks/package.json
@@ -9,7 +9,7 @@
     "dev": "babel-node ./run.js dev",
     "start": "babel-node ./run.js start",
     "test": "jest --testMatch=**/test/unit/*.test.js --verbose",
-    "test:functional": "babel-node ./run.js test:functional",
+    "test:e2e": "babel-node ./run.js test:e2e",
     "import-event-store": "babel-node ./run.js import-event-store",
     "export-event-store": "babel-node ./run.js export-event-store",
     "resolve-module-admin": "resolve-module-admin"
@@ -22,8 +22,8 @@
     "jsonwebtoken": "8.5.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-bootstrap": "0.32.1",
-    "react-helmet": "6.0.0",
+    "react-bootstrap": "1.3.0",
+    "react-helmet": "6.1.0",
     "react-redux": "7.2.0",
     "react-router": "5.1.2",
     "react-router-dom": "5.1.2",
