+++ examples/shopping-list-with-redux-hooks/run.js
@@ -77,7 +77,7 @@ void (async () => {
         break
       }
 
-      case 'test:functional': {
+      case 'test:e2e': {
         const moduleAdmin = resolveModuleAdmin()
         const resolveConfig = merge(
           defaultResolveConfig,
